+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2024-08-18
+
 ### Added
 
 - `combine` sub-command for the `fun` tool which combines multiple input
@@ -54,7 +56,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First public release.
 
-[unreleased]: https://gitlab.com/tozd/go/fun/-/compare/v0.3.0...main
+[unreleased]: https://gitlab.com/tozd/go/fun/-/compare/v0.4.0...main
+[0.4.0]: https://gitlab.com/tozd/go/fun/-/compare/v0.3.0...v0.4.0
 [0.3.0]: https://gitlab.com/tozd/go/fun/-/compare/v0.2.2...v0.3.0
 [0.2.2]: https://gitlab.com/tozd/go/fun/-/compare/v0.2.1...v0.2.2
 [0.2.1]: https://gitlab.com/tozd/go/fun/-/compare/v0.2.0...v0.2.1